refactor(user): only use public API
- rename Controller to PublicUserAuthController
- delete unused admin function api
