--- src/main/java/de/fsr/mariokart_backend/user/controller/UserCreateController.java
@@ -1,48 +1,0 @@
-package de.fsr.mariokart_backend.user.controller;
-
-import java.util.List;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import de.fsr.mariokart_backend.exception.EntityNotFoundException;
-import de.fsr.mariokart_backend.user.model.dto.AuthenticationRequestDTO;
-import de.fsr.mariokart_backend.user.model.dto.AuthenticationResponseDTO;
-import de.fsr.mariokart_backend.user.model.dto.UserCreationDTO;
-import de.fsr.mariokart_backend.user.model.dto.UserTokenDTO;
-import de.fsr.mariokart_backend.user.service.AuthenticationService;
-import jakarta.validation.Valid;
-import lombok.AllArgsConstructor;
-
-@RestController
-@RequestMapping("/users")
-@AllArgsConstructor
-public class UserCreateController {
-
-    private final AuthenticationService authenticationService;
-
-    @PostMapping("/login")
-    public ResponseEntity<AuthenticationResponseDTO> login(@RequestBody @Valid AuthenticationRequestDTO request) {
-        try {
-            return ResponseEntity.ok().body(authenticationService.authenticateUser(request));
-        } catch (BadCredentialsException ex) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-    }
-
-    @ResponseBody
-    @PostMapping
-    public ResponseEntity<List<UserTokenDTO>> createUsers(@RequestBody List<UserCreationDTO> userCreations) {
-        try {
-            return ResponseEntity.ok(UserTokenDTO.fromUserTokenList(authenticationService.createUsers(userCreations)));
-        } catch (EntityNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        }
-    }
-}
