+++ src/main/java/de/fsr/mariokart_backend/user/controller/pub/PublicUserAuthController.java
@@ -1,0 +1,33 @@
+package de.fsr.mariokart_backend.user.controller.pub;
+
+import de.fsr.mariokart_backend.controller.annotation.ApiController;
+import de.fsr.mariokart_backend.controller.annotation.ApiType;
+import de.fsr.mariokart_backend.controller.annotation.ControllerType;
+import de.fsr.mariokart_backend.user.model.dto.AuthenticationRequestDTO;
+import de.fsr.mariokart_backend.user.model.dto.AuthenticationResponseDTO;
+import de.fsr.mariokart_backend.user.service.AuthenticationService;
+import jakarta.validation.Valid;
+import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@AllArgsConstructor
+@ApiController(apiType = ApiType.PUBLIC, controllerType = ControllerType.USER)
+public class PublicUserAuthController {
+
+    private final AuthenticationService authenticationService;
+
+    @PostMapping("/login")
+    public ResponseEntity<AuthenticationResponseDTO> login(@RequestBody @Valid AuthenticationRequestDTO request) {
+        try {
+            return ResponseEntity.ok().body(authenticationService.authenticateUser(request));
+        } catch (BadCredentialsException ex) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+    }
+}
